Add Class Price

+++ src/main/java/com/epam/brest/file/FileReader.java
@@ -1,9 +1,11 @@
 package com.epam.brest.file;
 
+import com.epam.brest.price.Price;
+
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.util.Map;
 
 public interface FileReader {
- Map<Integer, BigDecimal> readFileFromResources(String fileName) throws IOException;
+ Map<BigDecimal,BigDecimal> readFileFromResources(String fileName) throws IOException;
 }
